requirements.txt
Flask==2.2.5
openai==0.27.8
python-dotenv==1.0.0